feat(hashers): hasher structs
